+++ backOffice/src/main/java/com/auth/entity/Role.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_Roles;
+    private Long id_Roles;
     private String nom;
 
     @ManyToMany
